final changes - add equals for answer, move handlersExc from ControllerAdvice to classesWhereThrowsExc (deleted ControllerAdvice.class)
