+++ src/main/java/org/skypro/skyexam/examinerservice/controller/JavaQuestionController.java
@@ -1,9 +1,14 @@
 package org.skypro.skyexam.examinerservice.controller;
 
 import org.skypro.skyexam.examinerservice.domain.Question;
+import org.skypro.skyexam.examinerservice.service.ExamError;
 import org.skypro.skyexam.examinerservice.service.JavaQuestionService;
+import org.skypro.skyexam.examinerservice.service.NoSuchQuestionException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDateTime;
 import java.util.Collection;
 
 @RestController
@@ -20,6 +25,13 @@ public class JavaQuestionController {
         return questionService.getQuestions();
     }
 
+    @ExceptionHandler(NoSuchQuestionException.class)
+    public ResponseEntity<ExamError> handleNoSuchQuestionException(NoSuchQuestionException e) {
+        String code = "NOT_FOUND";
+        String message = String.format("%s %s", LocalDateTime.now(), e.getMessage());
+        ExamError examError = new ExamError(code, message);
+        return new ResponseEntity<ExamError>(examError, HttpStatus.NOT_FOUND);
+    }
     @GetMapping("/remove")
     @ResponseBody
     public Question removeQuestion(@RequestParam("question") String question, @RequestParam("answer") String answer) {
